Split monolythic reducer out into slice reducers
