+++ src/components/App.js
@@ -20,7 +20,7 @@ class App extends Component {
 }
 
 const mapStateToProps = state => ({
-  appName: state.appName
+  appName: state.common.appName
 });
 
 export default connect(mapStateToProps)(App);
