+++ src/components/Home/Home.js
@@ -32,7 +32,7 @@ class Home extends Component {
 }
 
 const mapStateToProps = state => ({
-  appName: state.appName
+  appName: state.common.appName
 });
 
 const mapDispatchToProps = dispatch => ({
