+++ src/reducers/auth.js
@@ -1,0 +1,3 @@
+export default (state = {}, action) => {
+  return state;
+}
