+++ src/reducers/common.js
@@ -1,0 +1,7 @@
+const initialState = {
+  appName: 'Medium Lite'
+};
+
+export default (state = initialState, action) => {
+  return state;
+}
