+++ src/store.js
@@ -1,26 +1,10 @@
 import { applyMiddleware, createStore } from 'redux';
 
 import { promiseMiddleware } from './middleware';
-
-const initialState = {
-  appName: 'Medium Lite',
-  articles: null
-};
-
-const reducer = (state = initialState, action) => {
-  switch (action.type) {
-    case 'HOME_PAGE_LOADED':
-      return {
-        ...state,
-        articles: action.payload.articles
-      };
-    default:
-      return state;
-  }
-};
+import appReducer from './reducers';
 
 const middleware = applyMiddleware(promiseMiddleware);
 
-const store = createStore(reducer, middleware);
+const store = createStore(appReducer, middleware);
 
 export default store;
